Add the `from_huey_ids` function to `TaskHistoryQuerySet`
