+++ tubesync/common/models/tasks.py
@@ -1,10 +1,14 @@
+import uuid
 from datetime import timedelta
+from itertools import islice
 
-from django.db import models
+from django.db import connection, models, transaction
 from django.utils import timezone
 
 from ..json import JSONEncoder
 # from common.json import JSONEncoder
+from ..utils import is_empty_iterator
+#from common.utils import is_empty_iterator
 
 # cls = TaskHistory
 # TaskHistory is defined below this function in this file
@@ -43,8 +47,78 @@ def th_schedule(cls, task_wrapper, /, *args, remove_duplicates=False, vn_args=()
     task_wrapper.huey.enqueue(task_obj)
     return True
 
+# self = TaskHistoryQuerySet
+# TaskHistoryQuerySet is defined below this function in this file
+def thqs_from_huey_ids(self, /, huey_task_ids):
+    """
+    Robustly matches Huey task.id values to TaskHistory records.
+
+    Optimized for large datasets across SQLite, PostgreSQL, and MariaDB.
+    Bypasses SQL variable limits by using request-cycle temporary tables.
+    """
+    # 1. Guard clause for empty input - returns a valid, empty QuerySet
+    empty, huey_task_ids = is_empty_iterator(huey_task_ids)
+    if empty:
+        return self.none()
+
+    # 2. Dynamic metadata and unique naming
+    task_history_table = self.model._meta.db_table
+    unique_suffix = uuid.uuid4().hex[:8]
+    input_tmp = f"tmp_huey_ids_{unique_suffix}"
+    results_tmp = f"tmp_history_pks_{unique_suffix}"
+
+    # 3. Database operations
+    def validated_id_generator():
+        for tid in huey_task_ids:
+            try:
+                # Validates format and normalizes to lowercase dashed string
+                yield (str(uuid.UUID(str(tid))).lower(),)
+            except (ValueError, TypeError, AttributeError):
+                # Skip malformed IDs and log for troubleshooting
+                from common.logger import log
+                log.warning(f"Skipping malformed Huey task ID: {tid}")
+                continue
+
+    with transaction.atomic():
+        with connection.cursor() as cursor:
+            # Stage 1: Store and normalize input IDs
+            cursor.execute(f"CREATE TEMPORARY TABLE {input_tmp} (tid VARCHAR(40) PRIMARY KEY)")
+
+            # Single iteration: ensures dashed strings and lowercase for case-sensitive DBs
+            # Batch stream
+            batch_size = 40_000
+            stream = validated_id_generator()
+            while batch := list(islice(stream, batch_size)):
+                cursor.executemany(f"INSERT INTO {input_tmp} VALUES (%s)", batch)
+
+            # Stage 2: Filter to a PK-only results table
+            cursor.execute(f"CREATE TEMPORARY TABLE {results_tmp} (id BIGINT)")
+            cursor.execute(f"""
+                INSERT INTO {results_tmp} (id)
+                SELECT id FROM {task_history_table}
+                WHERE task_id IN (SELECT tid FROM {input_tmp})
+            """)
+
+            # Index the result PKs to ensure dashboard pagination and ordering are fast
+            cursor.execute(f"CREATE INDEX idx_{results_tmp} ON {results_tmp}(id)")
+
+            # Immediate cleanup of the input string table to save memory
+            cursor.execute(f"DROP TABLE IF EXISTS {input_tmp}")
+
+    # 4. Return a Lazy QuerySet via RawSQL to bypass RawQuerySet.clone() limitations.
+    # The 'results_tmp' table persists until the connection closes after the request.
+    return self.filter(
+        id__in=models.expressions.RawSQL(
+            f"SELECT id FROM {results_tmp}",
+            [],
+        )
+    )
+
 
 class TaskHistoryQuerySet(models.QuerySet):
+
+    def from_huey_ids(self, /, huey_task_ids):
+        return thqs_from_huey_ids(self, huey_task_ids)
 
     def running(self, now=None, within=None):
         if now is None:
@@ -135,7 +209,7 @@ class TaskHistory(models.Model):
         self.queue = self.queue or None
         self.verbose_name = self.verbose_name or None
         return super().save(*args, **kwargs)
-        
+
     def has_error(self):
         """
         Check if the last_error field is empty.
